pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "equilibrium"
version = "0.1.0"
description = ""
authors = ["Niklas Rosenstein <rosensteinniklas@gmail.com>"]
license = "MIT"
packages = [{ include = "equilibrium", from = "src" }]
classifiers = []
keywords = []

[tool.poetry.urls]
# "Bug Tracker" = ""
# Documentation = ""
# Homepage = ""
# Repository = ""

[tool.poetry.dependencies]
python = "^3.10"
"databind.json" = "^4.2.5"
pyyaml = "^6.0"
"nr.proxy" = "^1.1.1"
types-pyyaml = "^6.0.12.9"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"
isort = "*"
mypy = "*"
pytest = "*"

[tool.slap]
typed = true

[tool.slap.test]
check = "slap check"
mypy = "dmypy run src/"
pytest = "pytest src/ -vv"
black = "black --check src/"
isort = "isort --check-only src/"
flake8 = "flake8 src/"

[tool.slap.run]
fmt = "black src/ && isort src/"

[tool.mypy]
python_version = "3.10"
explicit_package_bases = true
mypy_path = ["src"]
namespace_packages = true
pretty = true
show_error_codes = true
show_error_context = true
strict = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[tool.isort]
profile = "black"
line_length = 120
combine_as_imports = true

[tool.black]
line-length = 120
